+++ finagle-thrift/src/main/scala/com/twitter/finagle/Thrift.scala
@@ -255,9 +255,6 @@ object Thrift
     protected type Out = Array[Byte]
     protected type Context = TransportContext
 
-    @deprecated("Use clientParam.protocolFactory", "2017-08-16")
-    protected def protocolFactory: TProtocolFactory = clientParam.protocolFactory
-
     @deprecated("Use clientParam.clientStats", "2017-08-16")
     override protected def stats: StatsReceiver = clientParam.clientStats
 
@@ -483,9 +480,6 @@ object Thrift
 
     @deprecated("Use serverParam.serverStats", "2017-08-16")
     override protected def serverStats: StatsReceiver = serverParam.serverStats
-
-    @deprecated("Use serverParam.protocolFactory", "2017-08-16")
-    protected def protocolFactory: TProtocolFactory = serverParam.protocolFactory
 
     @deprecated("Use serverParam.maxThriftBufferSize", "2017-08-16")
     override protected def maxThriftBufferSize: Int = serverParam.maxThriftBufferSize